+++ Assets/Scripts/Saving/PlayerWInventory.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 25ef36415791b25409f2eb39cc67b8f7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
