+++ Assets/Scripts/Saving/SaveLoadManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 97b12699b8b8e164aab90ee527488ccd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
