+++ Assets/Scripts/Saving/TempScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 22190e634f4c8524e9cee0d19c7e4479
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
